Flytt utils, flytt useScenarioDrawer, flytt felleskomponenter
