+++ plugins/ros/src/components/common/Spinner.tsx
@@ -1,6 +1,6 @@
 import { CircularProgress } from '@material-ui/core';
 import React from 'react';
-import { useSpinnerStyles } from './style';
+import { useSpinnerStyles } from '../../utils/style';
 
 export const Spinner = ({ size }: { size?: string | number }) => {
   const { container, spinner } = useSpinnerStyles();
