+++ plugins/ros/src/components/riScPlugin/RiScPlugin.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { Button, Grid, LinearProgress, Typography } from '@material-ui/core';
 import { ContentHeader, SupportButton } from '@backstage/core-components';
-import { useFetchRiScs, useScenarioDrawer } from '../utils/hooks';
+import { useFetchRiScs } from '../../utils/hooks';
 import { useParams } from 'react-router';
 import { Route, Routes, useSearchParams } from 'react-router-dom';
 import { riScRouteRef, scenarioRouteRef } from '../../routes';
@@ -12,17 +12,17 @@ import {
 import { ScenarioDrawer } from '../scenarioDrawer/ScenarioDrawer';
 import { RiskMatrix } from '../riskMatrix/RiskMatrix';
 import { ScenarioTable } from '../scenarioTable/ScenarioTable';
-import { ScenarioContext } from './ScenarioContext';
 import Alert from '@mui/material/Alert';
-import { getAlertSeverity } from '../utils/utilityfunctions';
+import { getAlertSeverity } from '../../utils/utilityfunctions';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
-import { pluginRiScTranslationRef } from '../utils/translations';
-import { Dropdown } from '../utils/Dropdown';
+import { pluginRiScTranslationRef } from '../../utils/translations';
 import { RiScDialog, RiScDialogStates } from '../riScDialog/RiScDialog';
 import { RiScInfo } from '../riScInfo/RiScInfo';
 import AddCircle from '@material-ui/icons/AddCircle';
-import { Spinner } from '../utils/Spinner';
 import { useLinearProgressStyle } from './linearProgressStyle';
+import { Spinner } from '../common/Spinner';
+import { Dropdown } from '../common/Dropdown';
+import { ScenarioProvider } from '../../ScenarioContext';
 
 const Plugin = () => {
   const params = useParams();
@@ -52,12 +52,6 @@ const Plugin = () => {
     isRequesting,
   } = useFetchRiScs(params.riScId);
 
-  const scenario = useScenarioDrawer(
-    selectedRiSc ?? null,
-    updateRiSc,
-    params.scenarioId,
-  );
-
   const [searchParams] = useSearchParams();
   const scenarioWizardStep =
     (searchParams.get('step') as ScenarioWizardSteps | null) || null;
@@ -67,99 +61,101 @@ const Plugin = () => {
   }, [resetRiScStatus, scenarioWizardStep]);
 
   return (
-    <>
-      <ScenarioContext.Provider value={scenario}>
-        {response && (
-          <Alert severity={getAlertSeverity(response.status)}>
-            <Typography>{response.statusMessage}</Typography>
-          </Alert>
-        )}
-        {isRequesting && <LinearProgress className={linearProgress} />}
+    <ScenarioProvider
+      riSc={selectedRiSc ?? null}
+      updateRiSc={updateRiSc}
+      scenarioIdFromParams={params.scenarioId}
+    >
+      {response && (
+        <Alert severity={getAlertSeverity(response.status)}>
+          <Typography>{response.statusMessage}</Typography>
+        </Alert>
+      )}
+      {isRequesting && <LinearProgress className={linearProgress} />}
 
-        {scenarioWizardStep !== null ? (
-          <ScenarioWizard
-            step={scenarioWizardStep}
-            isFetching={isFetching}
-            updateStatus={updateRiScStatus}
-          />
-        ) : (
-          <>
-            <ContentHeader title={t('contentHeader.title')}>
-              <SupportButton />
-            </ContentHeader>
+      {scenarioWizardStep !== null ? (
+        <ScenarioWizard
+          step={scenarioWizardStep}
+          isFetching={isFetching}
+          updateStatus={updateRiScStatus}
+        />
+      ) : (
+        <>
+          <ContentHeader title={t('contentHeader.title')}>
+            <SupportButton />
+          </ContentHeader>
 
-            {isFetching && <Spinner size={80} />}
+          {isFetching && <Spinner size={80} />}
 
-            <Grid container spacing={4}>
-              {riScs !== null && riScs.length !== 0 && (
-                <Grid
-                  item
-                  xs={12}
-                  sm={6}
+          <Grid container spacing={4}>
+            {riScs !== null && riScs.length !== 0 && (
+              <Grid
+                item
+                xs={12}
+                sm={6}
+                style={{
+                  maxWidth: '600px',
+                  minWidth: '300px',
+                }}
+              >
+                <Dropdown<string>
+                  options={riScs.map(riSc => riSc.content.title) ?? []}
+                  selectedValues={selectedRiSc?.content.title ?? ''}
+                  handleChange={title => selectRiSc(title)}
+                  variant="standard"
+                />
+              </Grid>
+            )}
+
+            {!isFetching && (
+              <Grid item xs>
+                <Button
+                  startIcon={<AddCircle />}
+                  variant="text"
+                  color="primary"
+                  onClick={openCreateRiScDialog}
                   style={{
-                    maxWidth: '600px',
-                    minWidth: '300px',
+                    minWidth: '205px',
                   }}
                 >
-                  <Dropdown<string>
-                    options={riScs.map(riSc => riSc.content.title) ?? []}
-                    selectedValues={selectedRiSc?.content.title ?? ''}
-                    handleChange={title => selectRiSc(title)}
-                    variant="standard"
+                  {t('contentHeader.createNewButton')}
+                </Button>
+              </Grid>
+            )}
+
+            {selectedRiSc && (
+              <>
+                <Grid item xs={12}>
+                  <RiScInfo
+                    riSc={selectedRiSc}
+                    approveRiSc={approveRiSc}
+                    edit={openEditRiScDialog}
                   />
                 </Grid>
-              )}
-
-              {!isFetching && (
-                <Grid item xs>
-                  <Button
-                    startIcon={<AddCircle />}
-                    variant="text"
-                    color="primary"
-                    onClick={openCreateRiScDialog}
-                    style={{
-                      minWidth: '205px',
-                    }}
-                  >
-                    {t('contentHeader.createNewButton')}
-                  </Button>
+                <Grid item xs md={7} lg={8}>
+                  <ScenarioTable riSc={selectedRiSc.content} />
                 </Grid>
-              )}
+                <Grid item xs md={5} lg={4}>
+                  <RiskMatrix riSc={selectedRiSc.content} />
+                </Grid>
+              </>
+            )}
+          </Grid>
+        </>
+      )}
 
-              {selectedRiSc && (
-                <>
-                  <Grid item xs={12}>
-                    <RiScInfo
-                      riSc={selectedRiSc}
-                      approveRiSc={approveRiSc}
-                      edit={openEditRiScDialog}
-                    />
-                  </Grid>
-                  <Grid item xs md={7} lg={8}>
-                    <ScenarioTable riSc={selectedRiSc.content} />
-                  </Grid>
-                  <Grid item xs md={5} lg={4}>
-                    <RiskMatrix riSc={selectedRiSc.content} />
-                  </Grid>
-                </>
-              )}
-            </Grid>
-          </>
-        )}
+      {RiScDialogState !== RiScDialogStates.Closed && (
+        <RiScDialog
+          onClose={closeRiScDialog}
+          createNewRiSc={createNewRiSc}
+          updateRiSc={updateRiSc}
+          riSc={selectedRiSc}
+          dialogState={RiScDialogState}
+        />
+      )}
 
-        {RiScDialogState !== RiScDialogStates.Closed && (
-          <RiScDialog
-            onClose={closeRiScDialog}
-            createNewRiSc={createNewRiSc}
-            updateRiSc={updateRiSc}
-            riSc={selectedRiSc}
-            dialogState={RiScDialogState}
-          />
-        )}
-
-        {!scenarioWizardStep && <ScenarioDrawer />}
-      </ScenarioContext.Provider>
-    </>
+      {!scenarioWizardStep && <ScenarioDrawer />}
+    </ScenarioProvider>
   );
 };
