--- plugins/ros/src/components/riScPlugin/ScenarioContext.tsx
@@ -1,4 +1,0 @@
-import { createContext } from 'react';
-import { ScenarioDrawerProps } from '../utils/hooks';
-
-export const ScenarioContext = createContext<ScenarioDrawerProps | null>(null);
