+++ plugins/ros/src/components/riskMatrix/AggregatedCost.tsx
@@ -1,12 +1,12 @@
 import React, { useState } from 'react';
 import { Box, IconButton, Typography } from '@material-ui/core';
 import InfoIcon from '@mui/icons-material/Info';
-import { RiSc } from '../utils/types';
+import { RiSc } from '../../utils/types';
 import { EstimatedRiskInfoDialog } from './EstimatedRiskInfoDialog';
-import { formatNumber } from '../utils/utilityfunctions';
-import { pluginRiScTranslationRef } from '../utils/translations';
+import { formatNumber } from '../../utils/utilityfunctions';
+import { pluginRiScTranslationRef } from '../../utils/translations';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
-import { useFontStyles } from '../utils/style';
+import { useFontStyles } from '../../utils/style';
 import { useAggregatedCostStyles } from './aggregatedCostStyle';
 
 interface AggregatedCostProps {
