+++ plugins/ros/src/components/scenarioDrawer/components/DeleteConfirmation.tsx
@@ -1,13 +1,13 @@
 import { Button, Dialog, DialogActions, DialogTitle } from '@material-ui/core';
-import React, { useContext } from 'react';
-import { ScenarioContext } from '../../riScPlugin/ScenarioContext';
-import { pluginRiScTranslationRef } from '../../utils/translations';
+import React from 'react';
+import { pluginRiScTranslationRef } from '../../../utils/translations';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
+import { useScenario } from '../../../ScenarioContext';
 
 export const DeleteConfirmation = () => {
   const { t } = useTranslationRef(pluginRiScTranslationRef);
   const { deleteConfirmationIsOpen, abortDeletion, confirmDeletion } =
-    useContext(ScenarioContext)!!;
+    useScenario();
 
   return (
     <Dialog open={deleteConfirmationIsOpen}>
