+++ plugins/ros/src/utils/hooks.ts
@@ -9,24 +9,16 @@ import {
   useRouteRef,
 } from '@backstage/core-plugin-api';
 import {
-  Action,
   ContentStatus,
   GithubRepoInfo,
   ProcessingStatus,
   RiSc,
   RiScStatus,
   RiScWithMetadata,
-  Risk,
-  Scenario,
   SubmitResponseObject,
 } from './types';
-import {
-  emptyAction,
-  emptyScenario,
-  requiresNewApproval,
-} from './utilityfunctions';
-import { consequenceOptions, probabilityOptions } from './constants';
-import { riScRouteRef, scenarioRouteRef } from '../../routes';
+import { requiresNewApproval } from './utilityfunctions';
+import { riScRouteRef } from '../routes';
 import { useLocation, useNavigate } from 'react-router';
 import {
   dtoToRiSc,
@@ -38,8 +30,6 @@ import {
   riScToDTOString,
 } from './DTOs';
 import { useEffectOnce } from 'react-use';
-import { useSearchParams } from 'react-router-dom';
-import { ScenarioWizardSteps } from '../scenarioWizard/ScenarioWizard';
 
 const useGithubRepositoryInformation = (): GithubRepoInfo => {
   const [, org, repo] =
@@ -224,381 +214,11 @@ const useFetch = () => {
   };
 };
 
-export interface ScenarioDrawerProps {
-  scenarioDrawerState: ScenarioDrawerState;
-
-  scenario: Scenario;
-  originalScenario: Scenario;
-  newScenario: () => void;
-  saveScenario: () => boolean;
-  editScenario: (step: ScenarioWizardSteps) => void;
-  isNewScenario: boolean;
-
-  openScenario: (id: string) => void;
-  closeScenario: () => void;
-
-  scenarioErrors: ScenarioErrors;
-  validateScenario: () => boolean;
-
-  deleteConfirmationIsOpen: boolean;
-  openDeleteConfirmation: () => void;
-  abortDeletion: () => void;
-  confirmDeletion: () => void;
-
-  setTitle: (title: string) => void;
-  setDescription: (description: string) => void;
-  setThreatActors: (threatActors: string[]) => void;
-  setVulnerabilities: (vulnerabilities: string[]) => void;
-  setProbability: (probabilityLevel: number) => void;
-  setConsequence: (consequenceLevel: number) => void;
-  setExistingActions: (existingActions: string) => void;
-  addAction: () => void;
-  updateAction: (action: Action) => void;
-  deleteAction: (action: Action) => void;
-  updateRemainingRisk: (remainingRisk: Risk) => void;
-  setRemainingProbability: (probabilityLevel: number) => void;
-  setRemainingConsequence: (consequenceLevel: number) => void;
-  setProbabilityAndRemainingProbability: (probabilityLevel: number) => void;
-  setConsequenceAndRemainingConsequence: (consequenceLevel: number) => void;
-
-  setFormError: (key: string) => void;
-  removeFormError: (key: string) => void;
-  hasFormErrors: () => boolean;
-  formFieldHasErrors: (key: string) => boolean;
-}
-
 export enum ScenarioDrawerState {
   Closed,
   Edit,
   View,
 }
-
-type ScenarioErrors = {
-  title: boolean;
-};
-
-const emptyScenarioErrors = (): ScenarioErrors => ({
-  title: false,
-});
-
-export const useScenarioDrawer = (
-  riSc: RiScWithMetadata | null,
-  updateRiSc: (riSc: RiSc) => void,
-  scenarioIdFromParams?: string,
-): ScenarioDrawerProps => {
-  // STATES
-  const [scenarioDrawerState, setScenarioDrawerState] = useState(
-    ScenarioDrawerState.Closed,
-  );
-
-  const [isNewScenario, setIsNewScenario] = useState(false);
-  const [formErrors, _setFormErrors] = useState<{ [key: string]: boolean }>({});
-  const [scenario, setScenario] = useState(emptyScenario());
-  const [originalScenario, setOriginalScenario] = useState(emptyScenario());
-  const [deleteConfirmationIsOpen, setDeleteConfirmationIsOpen] =
-    useState(false);
-  const [, setSearchParams] = useSearchParams();
-
-  const [scenarioErrors, setScenarioErrors] = useState(emptyScenarioErrors());
-  const navigate = useNavigate();
-  const getScenarioPath = useRouteRef(scenarioRouteRef);
-  const getRiScPath = useRouteRef(riScRouteRef);
-
-  // Open scenario when url changes
-  useEffect(() => {
-    if (riSc) {
-      // If there is no scenario ID in the URL, close the drawer and reset the scenario to an empty state
-      if (!scenarioIdFromParams) {
-        setScenarioDrawerState(ScenarioDrawerState.Closed);
-        const s = emptyScenario();
-        setScenario(s);
-        setOriginalScenario(s);
-        setScenarioErrors(emptyScenarioErrors());
-        return;
-      }
-
-      if (isNewScenario) {
-        // If this is a new scenario we should not set a "not found" state on path.
-        return;
-      }
-
-      const selectedScenario = riSc.content.scenarios.find(
-        s => s.ID === scenarioIdFromParams,
-      );
-
-      // If there is an invalid scenario ID in the URL, navigate to the RiSc with error state
-      if (!selectedScenario) {
-        navigate(getRiScPath({ riScId: riSc.id }), {
-          state: 'Risikoscenarioet du prøver å åpne eksisterer ikke',
-        });
-        return;
-      }
-
-      setScenario(selectedScenario);
-      setOriginalScenario(selectedScenario);
-      setScenarioDrawerState(ScenarioDrawerState.View);
-    }
-  }, [riSc, scenarioIdFromParams, getRiScPath, navigate, isNewScenario]);
-
-  // SCENARIO DRAWER FUNCTIONS
-  const openScenario = (id: string) => {
-    if (riSc) {
-      navigate(getScenarioPath({ riScId: riSc.id, scenarioId: id }));
-    }
-  };
-
-  // openNewScenarioWizard is used when opening the wizard with a new scenario. It explisit sets the searchParam to step=scenario.
-  const openNewScenarioWizard = (id: string, step: ScenarioWizardSteps) => {
-    if (riSc) {
-      navigate(
-        `${getScenarioPath({ riScId: riSc.id, scenarioId: id })}?step=${step}`,
-      );
-    }
-  };
-
-  const closeScenario = () => {
-    if (riSc) {
-      setIsNewScenario(false);
-      navigate(getRiScPath({ riScId: riSc.id }));
-    }
-  };
-
-  const editScenario = (step: ScenarioWizardSteps) => {
-    setScenarioDrawerState(ScenarioDrawerState.Closed);
-    setSearchParams({ step: step });
-  };
-
-  const validateScenario = () => {
-    if (scenario.title === '') {
-      setScenarioErrors({
-        ...scenarioErrors,
-        title: true,
-      });
-      return false;
-    }
-    return true;
-  };
-
-  const saveScenario = () => {
-    if (riSc) {
-      if (validateScenario()) {
-        const updatedScenarios = riSc.content.scenarios.some(
-          s => s.ID === scenario.ID,
-        )
-          ? riSc.content.scenarios.map(s =>
-              s.ID === scenario.ID ? scenario : s,
-            )
-          : riSc.content.scenarios.concat(scenario);
-        updateRiSc({ ...riSc.content, scenarios: updatedScenarios });
-        return true;
-      }
-    }
-    return false;
-  };
-
-  const openDeleteConfirmation = () => setDeleteConfirmationIsOpen(true);
-
-  const abortDeletion = () => setDeleteConfirmationIsOpen(false);
-
-  const confirmDeletion = () => {
-    if (riSc) {
-      setDeleteConfirmationIsOpen(false);
-      closeScenario();
-      const updatedScenarios = riSc.content.scenarios.filter(
-        s => s.ID !== scenario.ID,
-      );
-      updateRiSc({ ...riSc.content, scenarios: updatedScenarios });
-    }
-  };
-
-  const newScenario = () => {
-    if (riSc) {
-      setIsNewScenario(true);
-      const s = emptyScenario();
-      setScenario(s);
-      setOriginalScenario(s);
-      openNewScenarioWizard(s.ID, 'scenario');
-    }
-  };
-
-  // UPDATE SCENARIO FUNCTIONS
-
-  const setTitle = (title: string) => {
-    setScenarioErrors({
-      ...scenarioErrors,
-      title: false,
-    });
-    setScenario({
-      ...scenario,
-      title: title,
-    });
-  };
-
-  const setDescription = (description: string) => {
-    setScenario({
-      ...scenario,
-      description: description,
-    });
-  };
-
-  const setThreatActors = (threatActors: string[]) => {
-    setScenario({
-      ...scenario,
-      threatActors: threatActors,
-    });
-  };
-
-  const setVulnerabilities = (vulnerabilities: string[]) => {
-    setScenario({
-      ...scenario,
-      vulnerabilities: vulnerabilities,
-    });
-  };
-
-  const setProbability = (probabilityLevel: number) =>
-    setScenario({
-      ...scenario,
-      risk: {
-        ...scenario.risk,
-        probability: probabilityOptions[probabilityLevel - 1],
-      },
-    });
-
-  const setConsequence = (consequenceLevel: number) =>
-    setScenario({
-      ...scenario,
-      risk: {
-        ...scenario.risk,
-        consequence: consequenceOptions[consequenceLevel - 1],
-      },
-    });
-
-  const setExistingActions = (existingActions: string) => {
-    setScenario({
-      ...scenario,
-      existingActions: existingActions,
-    });
-  };
-
-  const addAction = () =>
-    setScenario({ ...scenario, actions: [...scenario.actions, emptyAction()] });
-
-  const updateAction = (action: Action) => {
-    const updatedAction = scenario.actions.some(a => a.ID === action.ID)
-      ? scenario.actions.map(a => (a.ID === action.ID ? action : a))
-      : [...scenario.actions, action];
-    setScenario({ ...scenario, actions: updatedAction });
-  };
-
-  const deleteAction = (action: Action) => {
-    const updatedAction = scenario.actions.filter(a => a.ID !== action.ID);
-    setScenario({ ...scenario, actions: updatedAction });
-  };
-
-  const updateRemainingRisk = (remainingRisk: Risk) =>
-    setScenario({ ...scenario, remainingRisk: remainingRisk });
-
-  const setRemainingProbability = (probabilityLevel: number) =>
-    setScenario({
-      ...scenario,
-      remainingRisk: {
-        ...scenario.remainingRisk,
-        probability: probabilityOptions[probabilityLevel - 1],
-      },
-    });
-
-  const setRemainingConsequence = (consequenceLevel: number) =>
-    setScenario({
-      ...scenario,
-      remainingRisk: {
-        ...scenario.remainingRisk,
-        consequence: consequenceOptions[consequenceLevel - 1],
-      },
-    });
-
-  const setProbabilityAndRemainingProbability = (probabilityLevel: number) =>
-    setScenario({
-      ...scenario,
-      risk: {
-        ...scenario.risk,
-        probability: probabilityOptions[probabilityLevel - 1],
-      },
-      remainingRisk: {
-        ...scenario.remainingRisk,
-        probability: probabilityOptions[probabilityLevel - 1],
-      },
-    });
-
-  const setConsequenceAndRemainingConsequence = (consequenceLevel: number) =>
-    setScenario({
-      ...scenario,
-      risk: {
-        ...scenario.risk,
-        consequence: consequenceOptions[consequenceLevel - 1],
-      },
-      remainingRisk: {
-        ...scenario.remainingRisk,
-        consequence: consequenceOptions[consequenceLevel - 1],
-      },
-    });
-
-  const setFormError = (key: string) => {
-    _setFormErrors({ ...formErrors, [key]: true });
-  };
-
-  const removeFormError = (key: string) => {
-    delete formErrors[key];
-    _setFormErrors({ ...formErrors });
-  };
-
-  const formFieldHasErrors = (key: string) =>
-    Object.keys(formErrors).includes(key);
-
-  const hasFormErrors = () => Object.keys(formErrors).length > 0;
-
-  return {
-    scenarioDrawerState,
-
-    scenario,
-    originalScenario,
-    newScenario,
-    saveScenario,
-    editScenario,
-    isNewScenario,
-
-    openScenario,
-    closeScenario,
-
-    scenarioErrors,
-    validateScenario,
-
-    deleteConfirmationIsOpen,
-    openDeleteConfirmation,
-    abortDeletion,
-    confirmDeletion,
-
-    setTitle,
-    setDescription,
-    setThreatActors,
-    setVulnerabilities,
-    setProbability,
-    setConsequence,
-    setExistingActions,
-    addAction,
-    updateAction,
-    deleteAction,
-    updateRemainingRisk,
-    setRemainingProbability,
-    setRemainingConsequence,
-    setProbabilityAndRemainingProbability,
-    setConsequenceAndRemainingConsequence,
-
-    setFormError,
-    hasFormErrors,
-    removeFormError,
-    formFieldHasErrors,
-  };
-};
 
 export const useFetchRiScs = (
   riScIdFromParams?: string,
